Changing api to separate file as an express function
