+++ package.json
@@ -17,6 +17,7 @@
     "express": "^4.21.2"
   },
   "devDependencies": {
-    "jest": "^29.7.0"
+    "jest": "^29.7.0",
+    "supertest": "^7.0.0"
   }
 }
